Adds MountSettings
With MountSettings a user is able to specify a list of known
usernames/hosts configurations with ports and private ssh keys.

The settings are stored as a [MountSettings] value in a
~/.vadosettings file.

If the file is absent or vado has failed to read it then the settings
are defaulted to standart vagrant ones.