Adds a 'vamount' command
Vamount is a command for easily mounting sshfs systems based on
~/.vadosettings configuration file.

Usage vamount [ssh options] remote_path [profile #]

The remote_path from the remote server specified
in the ~/.vadosettings file under number [profile #]
will be mounted in the current directory using sshfs

The ssh options must start with a dash '-'.
The profile number count starts from 1.
If the [profile #] is absent or is 0 then
the default configuration will be used.